Feat: 정리

+++ 27433.cpp
@@ -2,20 +2,20 @@
 
 using namespace std;
 
-int factorial(int number);
+long long factorial(int number);
 
 int main(){
     int num;
 
     cin >> num;
-    int result = factorial(num);
+    long long result = factorial(num);
     cout << result << endl;
 
     return 0;
 }
 
-int factorial(int number){
-    int ret = 1;
+long long factorial(int number){
+    long long ret = 1;
 
     for(int i=number; i>0; i--){
         ret = ret * i;
